refactor(domain): decompose exceptions.py into categorized package
Split 513 LOC monolithic exceptions.py file into a modular package:

- base.py (~144 LOC): BioETLError, CriticalError, RecoverableError, DataQualityError
- critical.py (~105 LOC): LockLostError, AuthFailureError, InfrastructureError, etc.
- recoverable.py (~115 LOC): RateLimitError, ApiError, TimeoutError, NetworkError, etc.
- data_quality.py (~69 LOC): SchemaViolationError, MissingRequiredFieldError, etc.
- storage.py (~50 LOC): TableNotFoundError, BucketNotFoundError, UploadError
- __init__.py (~80 LOC): Re-exports for backward compatibility

All files are now under 200 LOC. Backward compatibility preserved via
re-exports in __init__.py. Updated architecture test to scan package
structure.
